+++ package.json
@@ -87,7 +87,6 @@
     "@types/isomorphic-fetch": "^0.0.34",
     "@types/jest": "^23.3.2",
     "@types/node": "^10.11.0",
-    "@types/web3": "1.0.18",
     "@types/ws": "^6.0.1",
     "babel-eslint": "^10.0.1",
     "colors": "^1.3.2",
